requirements.txt
scrapy~=1.8.0
elasticsearch~=7.10.1
scrapy_proxies
python-dateutil
beautifulsoup4
scrapy-redis
pysummarization
scrapy-splash
socks
requests[socks]
pysocks
singleton~=0.1.0
DateTime~=4.2
Pillow~=8.3.1
Twisted~=19.10.0
ImageHash~=4.2.1
numpy~=1.21.1
scipy~=1.7.1
scikit-learn~=0.24.2
tqdm~=4.46.1
piexif~=1.1.3